wsd: sig: dump signal logs into files and extract later
Signals are now logged, safely, into signalLog files.
Each of wsd, frk, and kit processes have their own
unique signalLog files. During cleaning up, which
happens rather infrequently, wsd collects the
signalLog files, if any, and dumps them into the
main log file.

In addition, we now have USR2 signal that dumps the
stack-trace of each process. This is useful for
capturing the state of misbehaving instances.

This patch does not change the behavior of USR1.
Specifically, unlike USR2, USR1 is not forwarded
from wsd to frk and the kits. Also unlike USR2,
USR1 dumps into stderr.

Now SEGV and co are captured in the main log,
thanks to this patch, and this is done in a
signal-safe fashion.

Change-Id: I1d82f678f30f430f627692cc42961b1928f69e11
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>